Move DictionaryExample outside of package for clarity.
